Filterbar in documentCloud
